fix: break event poller import cycle
