+++ src/lib/server/events.ts
@@ -1,5 +1,6 @@
 export * from './events/types.js';
 export * from './events/state.js';
+export * from './events/dispatcher.js';
 export * from './events/bus.js';
 export * from './events/poller.js';
 export * from './events/shutdown.js';
